added in fuctionality to delete goals to spreadsheet as well as cleaned up CSVHandlers
