+++ ScheduleMe/src/com/scheduleMe/ViewGoalsFrame.java
@@ -1,5 +1,8 @@
 package com.scheduleMe;
 
+import com.scheduleMe.utility.goalsCSVHandler.FinancialGoalsCSVHandler;
+import com.scheduleMe.utility.goalsCSVHandler.*;
+
 import javax.swing.*;
 import javax.swing.border.EmptyBorder;
 import java.awt.*;
@@ -160,7 +163,15 @@ public class ViewGoalsFrame extends JFrame implements ActionListener {
             new ViewGoalsFrame(currentUser);
         }
         if (e.getActionCommand().startsWith("delete_")) {
+            GoalsCSVHandler goalsCSVHandler = new FinancialGoalsCSVHandler();
             int index = Integer.parseInt(e.getActionCommand().substring(7));
+            if (goals.get(index).getType().getCategory().equals("Financial")){
+                try {
+                    goalsCSVHandler.performDelete(goals.get(index), currentUser);
+                } catch (IOException ex) {
+                    throw new RuntimeException(ex);
+                }
+            }
             goals.remove(index);
             dispose();
             new ViewGoalsFrame(currentUser);
