+++ ScheduleMe/src/com/scheduleMe/utility/goalsCSVHandler/GoalsCSVHandler.java
@@ -4,7 +4,6 @@ import com.scheduleMe.Goal;
 import com.scheduleMe.User;
 
 import java.io.IOException;
-import java.util.ArrayList;
 
 public abstract class GoalsCSVHandler {
 
@@ -14,10 +13,14 @@ public abstract class GoalsCSVHandler {
     public   void performWrite(Goal goal, User user) throws IOException {
         goalsWriteBehavior.writeToCSV(goal, user);
     }
+    public void performDelete(Goal goal, User user) throws IOException {
+        goalsWriteBehavior.deleteFromCSV(goal, user);
+    }
 
     public   void performRead(User user) throws IOException {
         goalsReadBehavior.ReadFromCSV(user);
     }
+
 
     public GoalsWriteBehavior getGoalsWriteBehavior() {
         return goalsWriteBehavior;
