+++ ScheduleMe/src/com/scheduleMe/utility/goalsCSVHandler/RelationshipGoalsCSVHandler.java
@@ -2,7 +2,7 @@ package com.scheduleMe.utility.goalsCSVHandler;
 
 public class RelationshipGoalsCSVHandler extends GoalsCSVHandler{
     public RelationshipGoalsCSVHandler(){
-        goalsWriteBehavior = new WriteRelationshipGoal();
+       // goalsWriteBehavior = new WriteRelationshipGoal();
         goalsReadBehavior = new ReadRelationshipGoal();
     }
 }
