+++ ScheduleMe/src/com/scheduleMe/utility/goalsCSVHandler/WritePersonalGoal.java
@@ -1,0 +1,20 @@
+package com.scheduleMe.utility.goalsCSVHandler;
+
+import com.scheduleMe.Goal;
+import com.scheduleMe.User;
+
+import java.io.IOException;
+
+public class WritePersonalGoal implements GoalsWriteBehavior {
+
+    @Override
+    public void writeToCSV(Goal goal, User user) throws IOException {
+        System.out.println("Wrote Personal Goal!");
+
+    }
+
+    @Override
+    public void deleteFromCSV(Goal goal, User user) throws IOException {
+
+    }
+}
